fix: 검색 서버 컴포넌트 분리
